requirements.txt
hydra-core
pyrootutils
diffusers>=0.30.3
peft
timm
fairscale
flask
torchdata
kornia==0.7.3
braceexpand
av
decord==0.6.0
accelerate>=0.34.2
protobuf
opencv-python
deepspeed
transformers>=4.44.2
numpy==2.0.1
sentencepiece>=0.2.0
SwissArmyTransformer>=0.4.12
gradio>=4.44.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
openai>=1.45.0
moviepy>=1.0.3
pillow==9.5.0
scikit-video
safetensors
tokenizers==0.19
xformers
torch==2.4.0
torchvision==0.19.0
wandb==0.17.5
pytorch_lightning==2.3.3
open_clip_torch==2.22.0
beartype==0.18.5
fsspec==2024.5.0
scipy==1.14.0
loguru
SwissArmyTransformer==0.4.12
omegaconf==2.3.0